Fim videos autorizacao

+++ app/Http/Controllers/ProjectController.php
@@ -7,6 +7,8 @@ use CodeProject\Services\ProjectService;
 use Illuminate\Http\Request;
 use CodeProject\Http\Requests;
 
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
+
 class ProjectController extends Controller
 {
 
@@ -30,7 +32,7 @@ class ProjectController extends Controller
     
     public function index() 
     {
-    	return $this->repository->all();
+    	return $this->repository->findWhere(['owner_id' => Authorizer::getResourceOwnerId()]);
     }
     
     public function store(Request $request)
@@ -40,16 +42,31 @@ class ProjectController extends Controller
 
     public function update(Request $request, $id)
     {
+        if ($this->checkProjectPermissions($id) == false)
+        {
+            return response()->json([ 'error' => 'Access forbidden' ]);
+        }
+
         return $this->service->update($request->all(), $id);
     }
     
     public function show($id)
     {
+        if ($this->checkProjectPermissions($id) == false)
+        {
+            return response()->json([ 'error' => 'Access forbidden' ]);
+        }
+
     	return $this->repository->find($id);
     }
     
     public function destroy($id)
     {
+        if ($this->checkProjectPermissions($id) == false)
+        {
+            return response()->json([ 'error' => 'Access forbidden' ]);
+        }
+
 		try
 	    {
 	    	$this->repository->find($id)->delete();
@@ -61,4 +78,27 @@ class ProjectController extends Controller
 	    }    	
 	}
 
+    private function checkProjectOwner($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) == true or 
+            $this->checkProjectMember($projectId) == true)
+        {
+            return true;
+        }
+        return false;
+    }
 }
